% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_selection.R
\name{bracket_selection}
\alias{bracket_selection}
\alias{set_brackets}
\alias{get_brackets}
\title{Add customised brackets around selection}
\usage{
bracket_selection()

set_brackets(left = "{", right = "}")

get_brackets()
}
\arguments{
\item{left}{left of the bracket}

\item{right}{right of the bracket}
}
\value{
Return current brackets and print (\code{cat}) brackets around the word "example".
}
\description{
Add customised brackets around selection by setting addin using \code{bracket_selection}.
Set brackets and get currently set brackets.
Use two backslash "\\\\" when input backslash. See \code{Examples}.
}
\examples{
set_brackets(left = "\\\\boldsymbol{", right = "}")
}
\author{
Yangzhuoran Yang
}
