% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_purl_section.R
\name{insert_rmarkdown_chunk_editlabel}
\alias{insert_rmarkdown_chunk_editlabel}
\title{Insert Rmarkdown Chunk (Edit Label)}
\usage{
insert_rmarkdown_chunk_editlabel()
}
\description{
\code{insert_rmarkdown_chunk_editlabel} insert Rmarkdown Chunk with cursor placed at the label.
Just a shortcut to insert chunk without any code in it to use with \code{knitr::read_chunk}.
I suggest binding this addin with shortcut \code{Ctrl+Alt+O}.
}
\seealso{
\code{knitr::read_chunk}, \code{insert_purl_section}
}
\author{
Yangzhuoran Yang
}
