% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_run_operator.R
\name{insert_run_operator}
\alias{insert_run_operator}
\title{Insert LazyScript Run Operator}
\usage{
insert_run_operator()
}
\description{
\code{insert_run_operator} insert \code{\%run\%} at cursor. Binding to \code{Alt+,} is suggested.
}
\seealso{
\code{\%run\%}, \code{LazyScript}
}
\author{
Yangzhuoran Yang
}
