% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_scoping_assignment.R
\name{insert_scoping_assignment}
\alias{insert_scoping_assignment}
\title{Insert Scoping Assignment Operator}
\usage{
insert_scoping_assignment()
}
\description{
\code{insert_scoping_assignment} insert \code{<<-} at cursor. Binding to \code{Alt+=} is suggested.
}
\seealso{
\code{assignOps}
}
\author{
Yangzhuoran Yang
}
