% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazylm.R
\name{lazy.lm}
\alias{lazy.lm}
\title{Present a linear model in LaTex math format.}
\usage{
lazy.lm(model, digits = 2)
}
\arguments{
\item{model}{A linear regression model normally generated by function
\code{lm} or \code{glm}.}

\item{digits}{A number that defines the number of digits to containin
output.}
}
\value{
The code for the model presented in LaTex math format.
}
\description{
Present a linear model in LaTex math format.
}
\examples{
model <- lm(Murder ~ Assault + UrbanPop + Rape, USArrests)
lazy.lm(model)
}
