% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytype-package.R
\docType{package}
\name{lazytype-package}
\alias{lazytype-package}
\title{lazytype: Functions and Addins to save keystrokes and clicks}
\description{
The package provides addins for inserting code and running code in a different manner, and functions to allow interaction between scripts and between script and rmarkdown file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.yangzhuoranyang.com/lazytype/}
  \item \url{https://github.com/FinYang/lazytype}
  \item Report bugs at \url{https://github.com/FinYang/lazytype/issues}
}

}
\author{
\strong{Maintainer}: Yangzhuoran Yang \email{Fin.Yang@monash.edu} (\href{https://orcid.org/0000-0002-1232-8017}{ORCID})

}
\keyword{internal}
