% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminder.R
\name{reminder}
\alias{reminder}
\title{Show reminder every specified hours}
\usage{
reminder(
  hours = 1,
  message = "It's time to get up, drink some water and walk around.",
  dialog = TRUE,
  beep = 1,
  life = 5
)
}
\arguments{
\item{hours}{Numeric. Play a sound reminder every \code{hours} hours.}

\item{message}{The message to show in dialog.}

\item{dialog}{Logical. Whether to show a dialog remainder if reminder is set. Only support Windows system.}

\item{beep}{Numeric. The number of the sound to beep in \code{beepr::beep}. Do not beep if \code{NULL}.}

\item{life}{Numeric. The number of hours that you want to keep the reminder runing (the reminder's life).}
}
\description{
Show reminder every specified hours
}
\author{
Yangzhuoran Yang
}
