% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_to_cursor.R
\name{run_all_to_cursor}
\alias{run_all_to_cursor}
\title{Run All up to Cursor}
\usage{
run_all_to_cursor()
}
\description{
\code{run_all_to_cursor} by its name, run the code in the current script above the cursor.
I suggest binding this addin with shortcut \code{Ctrl+Alt+R}.
}
\seealso{
\code{run_selected_arguments}
}
\author{
Yangzhuoran Yang
}
