% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_current_purl_section.R
\name{run_current_purl_section}
\alias{run_current_purl_section}
\title{Run Current Purl Section}
\usage{
run_current_purl_section()
}
\description{
\code{run_current_purl_section} by its name, run the code in the purl section that the cursor is in.
I suggest binding this addin with shortcut \code{Shift+Alt+R}.
}
\seealso{
\code{run_selected_arguments}, \code{run_all_to_cursor}, \code{LazyScript}
}
\author{
Yangzhuoran Yang
}
