% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_selected_arguments.R
\name{run_selected_arguments}
\alias{run_selected_arguments}
\title{Run Selected Arguments}
\usage{
run_selected_arguments()
}
\description{
\code{run_selected_arguments} runs selected code, ignoring the comma.
It normally is used to set the default arguments of a function where the arguments are separated by comma.
Useful for testing function with different arguments.
I suggest binding this addin with shortcut \code{Alt+R}
}
\seealso{
\code{run_all_to_cursor}, \code{run_current_purl_section}
}
\author{
Yangzhuoran Yang
}
