% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazytype-package.R
\docType{package}
\name{lazytype}
\alias{lazytype}
\title{Functions and Addins to save keystrokes and clicks}
\description{
The lazytype package provides addins for inserting code and running code in a different manner,
and functions to interact scripts with each other and to interact script with rmarkdown file.
}
\details{
\url{http://pkg.yangzhuoranyang.com/lazytype/}
}
\author{
Yangzhuoran Yang
}
